refactored snapshot resources binding
