+++ envoy-control-core/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/protocol/HttpMethod.kt
@@ -1,4 +1,4 @@
-package pl.allegro.tech.servicemesh.envoycontrol.snapshot
+package pl.allegro.tech.servicemesh.envoycontrol.protocol
 
 enum class HttpMethod {
     GET, HEAD, POST, PUT, DELETE, OPTIONS, TRACE, CONNECT, PATCH
