+++ envoy-control-core/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/protocol/TlsUtils.kt
@@ -1,4 +1,4 @@
-package pl.allegro.tech.servicemesh.envoycontrol.snapshot
+package pl.allegro.tech.servicemesh.envoycontrol.protocol
 
 class TlsUtils {
     companion object {
