+++ envoy-control-core/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/snapshot/EnvoySnapshotFactory.kt
@@ -1,9 +1,5 @@
 package pl.allegro.tech.servicemesh.envoycontrol.snapshot
 
-import com.google.protobuf.ListValue
-import com.google.protobuf.Struct
-import com.google.protobuf.UInt32Value
-import com.google.protobuf.Value
 import com.google.protobuf.util.Durations
 import io.envoyproxy.controlplane.cache.Snapshot
 import io.envoyproxy.envoy.api.v2.Cluster
@@ -11,13 +7,6 @@ import io.envoyproxy.envoy.api.v2.ClusterLoadAssignment
 import io.envoyproxy.envoy.api.v2.Listener
 import io.envoyproxy.envoy.api.v2.RouteConfiguration
 import io.envoyproxy.envoy.api.v2.auth.Secret
-import io.envoyproxy.envoy.api.v2.core.Address
-import io.envoyproxy.envoy.api.v2.core.Locality
-import io.envoyproxy.envoy.api.v2.core.Metadata
-import io.envoyproxy.envoy.api.v2.core.SocketAddress
-import io.envoyproxy.envoy.api.v2.endpoint.Endpoint
-import io.envoyproxy.envoy.api.v2.endpoint.LbEndpoint
-import io.envoyproxy.envoy.api.v2.endpoint.LocalityLbEndpoints
 import io.micrometer.core.instrument.MeterRegistry
 import io.micrometer.core.instrument.Timer
 import pl.allegro.tech.servicemesh.envoycontrol.groups.AllServicesGroup
@@ -29,20 +18,20 @@ import pl.allegro.tech.servicemesh.envoycontrol.groups.ServicesGroup
 import pl.allegro.tech.servicemesh.envoycontrol.services.MultiClusterState
 import pl.allegro.tech.servicemesh.envoycontrol.services.ServiceInstance
 import pl.allegro.tech.servicemesh.envoycontrol.services.ServiceInstances
-import pl.allegro.tech.servicemesh.envoycontrol.snapshot.listeners.EnvoyListenersFactory
-import pl.allegro.tech.servicemesh.envoycontrol.snapshot.routing.ServiceTagMetadataGenerator
-import pl.allegro.tech.servicemesh.envoycontrol.services.Locality as LocalityEnum
+import pl.allegro.tech.servicemesh.envoycontrol.snapshot.resource.EnvoyClustersFactory
+import pl.allegro.tech.servicemesh.envoycontrol.snapshot.resource.EnvoyEgressRoutesFactory
+import pl.allegro.tech.servicemesh.envoycontrol.snapshot.resource.EnvoyEndpointsFactory
+import pl.allegro.tech.servicemesh.envoycontrol.snapshot.resource.EnvoyIngressRoutesFactory
+import pl.allegro.tech.servicemesh.envoycontrol.snapshot.resource.listeners.EnvoyListenersFactory
 
-internal class EnvoySnapshotFactory(
+class EnvoySnapshotFactory(
     private val ingressRoutesFactory: EnvoyIngressRoutesFactory,
     private val egressRoutesFactory: EnvoyEgressRoutesFactory,
     private val clustersFactory: EnvoyClustersFactory,
+    private val endpointsFactory: EnvoyEndpointsFactory,
     private val listenersFactory: EnvoyListenersFactory,
     private val snapshotsVersions: SnapshotsVersions,
     private val properties: SnapshotProperties,
-    private val serviceTagFilter: ServiceTagMetadataGenerator = ServiceTagMetadataGenerator(
-        properties.routing.serviceTags
-    ),
     private val defaultDependencySettings: DependencySettings =
         DependencySettings(
             handleInternalRedirect = properties.egress.handleInternalRedirect,
@@ -63,7 +52,7 @@ internal class EnvoySnapshotFactory(
         val clusters = clustersFactory.getClustersForServices(clusterConfigurations.values, communicationMode)
         val securedClusters = clustersFactory.getSecuredClusters(clusters)
 
-        val endpoints: List<ClusterLoadAssignment> = createLoadAssignment(
+        val endpoints: List<ClusterLoadAssignment> = endpointsFactory.createLoadAssignment(
             clusters = clusterConfigurations.keys,
             multiClusterState = servicesStates
         )
@@ -218,6 +207,7 @@ internal class EnvoySnapshotFactory(
         globalSnapshot: GlobalSnapshot
     ): Snapshot {
 
+        // TODO(dj): This is where serious refactoring needs to be done
         val egressRouteSpecification = getEgressRoutesSpecification(group, globalSnapshot)
 
         val clusters: List<Cluster> =
@@ -254,115 +244,6 @@ internal class EnvoySnapshotFactory(
         )
     }
 
-    private fun createEndpointsGroup(
-        serviceInstances: ServiceInstances,
-        zone: String,
-        priority: Int
-    ): LocalityLbEndpoints =
-        LocalityLbEndpoints.newBuilder()
-            .setLocality(Locality.newBuilder().setZone(zone).build())
-            .addAllLbEndpoints(serviceInstances.instances.map { createLbEndpoint(it, serviceInstances.serviceName) })
-            .setPriority(priority)
-            .build()
-
-    private fun createLbEndpoint(serviceInstance: ServiceInstance, serviceName: String): LbEndpoint {
-        return LbEndpoint.newBuilder()
-            .setEndpoint(
-                buildEndpoint(serviceInstance)
-            )
-            .setMetadata(serviceInstance, serviceName)
-            .setLoadBalancingWeightFromInstance(serviceInstance)
-            .build()
-    }
-
-    private fun buildEndpoint(serviceInstance: ServiceInstance): Endpoint.Builder {
-        return Endpoint.newBuilder()
-            .setAddress(
-                buildAddress(serviceInstance)
-            )
-    }
-
-    private fun buildAddress(serviceInstance: ServiceInstance): Address.Builder {
-        return Address.newBuilder()
-            .setSocketAddress(
-                buildSocketAddress(serviceInstance)
-            )
-    }
-
-    private fun buildSocketAddress(serviceInstance: ServiceInstance): SocketAddress.Builder {
-        return SocketAddress.newBuilder()
-            .setAddress(serviceInstance.address)
-            .setPortValue(serviceInstance.port)
-            .setProtocol(SocketAddress.Protocol.TCP)
-    }
-
-    private fun LbEndpoint.Builder.setMetadata(instance: ServiceInstance, serviceName: String): LbEndpoint.Builder {
-        val metadataKeys = Struct.newBuilder()
-
-        if (properties.loadBalancing.canary.enabled && instance.canary) {
-            metadataKeys.putFields(
-                properties.loadBalancing.canary.metadataKey,
-                Value.newBuilder().setStringValue(properties.loadBalancing.canary.headerValue).build()
-            )
-        }
-        if (instance.regular) {
-            metadataKeys.putFields(
-                properties.loadBalancing.regularMetadataKey,
-                Value.newBuilder().setBoolValue(true).build()
-            )
-        }
-
-        if (properties.routing.serviceTags.enabled) {
-            addServiceTagsToMetadata(metadataKeys, instance, serviceName)
-        }
-
-        return setMetadata(Metadata.newBuilder().putFilterMetadata("envoy.lb", metadataKeys.build()))
-    }
-
-    private fun addServiceTagsToMetadata(metadata: Struct.Builder, instance: ServiceInstance, serviceName: String) {
-        serviceTagFilter.getAllTagsForRouting(serviceName, instance.tags)?.let { tags ->
-            metadata.putFields(
-                properties.routing.serviceTags.metadataKey,
-                Value.newBuilder()
-                    .setListValue(ListValue.newBuilder()
-                        .addAllValues(tags.map { Value.newBuilder().setStringValue(it).build() }.asIterable())
-                    ).build()
-            )
-        }
-    }
-
-    private fun LbEndpoint.Builder.setLoadBalancingWeightFromInstance(instance: ServiceInstance): LbEndpoint.Builder =
-        when (properties.loadBalancing.weights.enabled) {
-            true -> setLoadBalancingWeight(UInt32Value.of(instance.weight))
-            false -> this
-        }
-
-    private fun toEnvoyPriority(locality: LocalityEnum): Int = if (locality == LocalityEnum.LOCAL) 0 else 1
-
-    private fun createLoadAssignment(
-        clusters: Set<String>,
-        multiClusterState: MultiClusterState
-    ): List<ClusterLoadAssignment> {
-
-        return clusters
-            .map { serviceName ->
-                val localityLbEndpoints = multiClusterState
-                    .mapNotNull {
-                        val locality = it.locality
-                        val cluster = it.cluster
-
-                        it.servicesState.get(serviceName)?.let {
-                            createEndpointsGroup(it, cluster, toEnvoyPriority(locality))
-                        }
-                    }
-
-                ClusterLoadAssignment.newBuilder()
-                    .setClusterName(serviceName)
-                    .addAllEndpoints(localityLbEndpoints)
-                    .build()
-            }
-    }
-
     private fun createSnapshot(
         clusters: List<Cluster> = emptyList(),
         clustersVersion: ClustersVersion = ClustersVersion.EMPTY_VERSION,
@@ -393,7 +274,7 @@ data class ClusterConfiguration(
     val mtlsEnabled: Boolean
 )
 
-internal class RouteSpecification(
+class RouteSpecification(
     val clusterName: String,
     val routeDomain: String,
     val settings: DependencySettings
