+++ envoy-control-core/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/snapshot/resource/AdminRoutesFactory.kt
@@ -1,4 +1,4 @@
-package pl.allegro.tech.servicemesh.envoycontrol.snapshot
+package pl.allegro.tech.servicemesh.envoycontrol.snapshot.resource
 
 import io.envoyproxy.envoy.api.v2.core.DataSource
 import io.envoyproxy.envoy.api.v2.route.DirectResponseAction
@@ -7,8 +7,10 @@ import io.envoyproxy.envoy.api.v2.route.RedirectAction
 import io.envoyproxy.envoy.api.v2.route.Route
 import io.envoyproxy.envoy.api.v2.route.RouteAction
 import io.envoyproxy.envoy.api.v2.route.RouteMatch
+import pl.allegro.tech.servicemesh.envoycontrol.protocol.HttpMethod
+import pl.allegro.tech.servicemesh.envoycontrol.snapshot.RoutesProperties
 
-internal class AdminRoutesFactory(
+class AdminRoutesFactory(
     private val properties: RoutesProperties
 ) {
     private val adminRouteAction = RouteAction.newBuilder()
