+++ envoy-control-core/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/snapshot/resource/RoutesMatchers.kt
@@ -1,6 +1,7 @@
-package pl.allegro.tech.servicemesh.envoycontrol.snapshot
+package pl.allegro.tech.servicemesh.envoycontrol.snapshot.resource
 
 import io.envoyproxy.envoy.api.v2.route.HeaderMatcher
+import pl.allegro.tech.servicemesh.envoycontrol.protocol.HttpMethod
 
 fun httpMethodMatcher(method: HttpMethod): HeaderMatcher = exactHeader(":method", method.name)
