+++ envoy-control-core/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/snapshot/resource/routing/ServiceTagFilter.kt
@@ -1,4 +1,4 @@
-package pl.allegro.tech.servicemesh.envoycontrol.snapshot.routing
+package pl.allegro.tech.servicemesh.envoycontrol.snapshot.resource.routing
 
 import io.envoyproxy.envoy.config.filter.http.header_to_metadata.v2.Config
